Create README.md
A simple Node.js project that allows you to manage a to-do list and interact with files in a specified folder. It provides endpoints for adding, viewing, and deleting to-do items, as well as fetching and displaying files and their contents in the given folder structure. The to-do data is stored in a JSON file, and all routes are accessible through an Express server.